+++ src/Campaign.php
@@ -117,6 +117,12 @@ class Campaign extends Plugin
         // Register tracker controller shorthand
         $this->controllerMap = ['t' => TrackerController::class];
 
+        // Console request
+        if (Craft::$app->getRequest()->getIsConsoleRequest()) {
+            // Add console commands
+            $this->controllerNamespace = __NAMESPACE__.'\console\controllers';
+        }
+
         // Register Twig extension
         Craft::$app->view->registerTwigExtension(new CampaignTwigExtension());
 
@@ -291,8 +297,10 @@ class Campaign extends Plugin
      * @param array $params
      * @param string|null $category
      */
-    public function logUserAction(string $message, array $params, string $category = 'Campaign')
+    public function logUserAction(string $message, array $params, string $category = null)
     {
+        $category = $category ?? 'Campaign';
+
         $params['username'] = Craft::$app->getUser()->getIdentity()->username;
 
         Craft::warning(Craft::t('campaign', $message, $params), $category);